requirements.txt
pillow
requests
PyYAML
